Do not call speaker.{Resume(),Suspend()} in sound.Play* anymore
This is causing more issues than it needed to, and the CPU usage while
it is unfortunate it is not that bad.

We will still call speaker.Suspend() when the sound is disabled and
speaker.Resume() when the sound is re-enabled. This results in 0% CPU
usage when this feature is not used, and doesn't need to bring back the
complicated initialisation code we had before.